Ash.Flow.StepHelpers (ash_flow v0.1.1)

Template functions available while declaring steps.

Functions

arg(name)

Accesses a flow argument

element(element)

Accesses the value being iterated over for a given map step. The name is required so that map steps can be nested.

expr(expr)
(macro)

Constructs an expression that can access

merge(results)

Accesses a flow argument

path(template, path)

Accesses a path in a value lazily. Supports the value being a template, e.g path(result(:foo), [:bar, 0, :baz])

range(start, finish)

Creates a range lazily. Supports the start or finish being a template, e.g range(result(:foo), result(:bar))

result(step)

Accesses the result of a step