Ash.Flow.Template (ash_flow v0.1.1)

Helpers for working with Ash.Flow templates.

At first glance it would seem that all of this could be replaced by some calls to Macro.prewalk but that is unfortunately not the case. We don't traverse through structs, and in some cases we do things like return ranges/merge maps that require having the entire nested part of the template, i.e {:_merge, [list, of, maps]} has to get the fully handled list of maps, and so has to call itself.

Functions

arg_refs(input)

element_refs(input)

handle_input_template(action_input, input)

is_template?(value)

remap_result_references(action_input, prefix)

result_refs(input)

set_dependent_values(action_input, input)